test(all): seperate browserstack task to config/browserstack.js
Ref #86
