+++ Gruntfile.js
@@ -1,6 +1,7 @@
 /*global module:false*/
-function getConfig(name){
-	return require("./config/"+name);
+
+function getConfig(name) {
+	return require("./config/" + name);
 }
 
 module.exports = function(grunt) {
@@ -8,8 +9,9 @@ module.exports = function(grunt) {
 	require("time-grunt")(grunt);
 	require("load-grunt-tasks")(grunt);
 	
-	var isBrowserStack = process.env.BROWSERSTACK_USERNAME && process.env.BROWSERSTACK_USERNAME;
-
+	// register browserstack task 
+	require("./config/browserstack.js")(grunt);
+	
 	var config = {
 		"pkg": grunt.file.readJSON("package.json"),
 		"clean": getConfig("clean"),
@@ -25,12 +27,11 @@ module.exports = function(grunt) {
 		"uglify": getConfig("uglify"),
 		"watch": getConfig("watch")
 	};
-
 	grunt.initConfig(config);
-
+	
 	grunt.registerTask("test", function() {
 		var eachfile = Array.prototype.slice.apply(arguments);
-		if(eachfile.length) {
+		if (eachfile.length) {
 			eachfile = eachfile.map(function(v) {
 				return "test/" + v + ".test.html";
 			}, this);
@@ -41,60 +42,7 @@ module.exports = function(grunt) {
 		grunt.log.oklns(grunt.config.get("qunit.each"));
 		grunt.task.run("qunit:each");
 	});
-
-	grunt.registerTask("browserstack", isBrowserStack ? function() {	
-		var eachfile = Array.prototype.slice.apply(arguments);
-		var taskList;
-		
-		if(eachfile.length) {
-			taskList = eachfile.map(function(v) {
-				return "browserstack_runner:" + v;
-			}, this);
-		} else {
-			var fs = require("fs");
-			
-			// fetch module name list from html files in test directory 
-			taskList = fs.readdirSync(
-				"test/"
-			).filter(function(val){
-				return val.split(".").length === 3 && val.split(".")[0] !== "buildMerge";
-			}).map(function(val){
-				return "browserstack_runner:" + val.split(".")[0];
-			});
-		}
-		grunt.task.run(taskList);
-	} : function() {
-		grunt.log.oklns(
-			"no BROWSER_STACK_USERNAME, BROWSER_STACK_ACCESS_KEY env"
-		);
-	});
-
-	grunt.registerTask("browserstack_runner", isBrowserStack ? function() {	
-		var fs = require("fs");
-		var exec = require("child_process").exec;
-		var targetModuleName = arguments[0];
-		var browserstackConfig = require("./config/browserstack.js").getConfig(targetModuleName);
-		var tempBrowserstackConfig = "config/browertstack.config.json";		
-		
-		fs.writeFileSync(tempBrowserstackConfig, JSON.stringify(browserstackConfig), "utf8");
-		process.env["BROWSERSTACK_JSON"] = tempBrowserstackConfig;
-	   
-	    var done = this.async();
-	    var subProcess = exec("node_modules/.bin/browserstack-runner", function (err) {
-			process.env["BROWSERSTACK_JSON"] = "";
-			fs.unlinkSync(tempBrowserstackConfig);
-			done(err ? false : true);
-	    });
-
-	    subProcess.stdout.on("data", function (_data) {
-		    grunt.log.writeln(_data.trim());
-	    });	    
-	} : function() {
-		grunt.log.oklns(
-			"no BROWSER_STACK_USERNAME, BROWSER_STACK_ACCESS_KEY env"
-		);
-	});
-
+	
 	grunt.registerTask("validate-commit", function() {
 		var fs = require("fs");
 		if (grunt.file.exists(".git/hooks/commit-msg")) {
@@ -103,7 +51,7 @@ module.exports = function(grunt) {
 		grunt.file.copy("config/validate-commit-msg.js", ".git/hooks/commit-msg", { force: true });
 		fs.chmodSync(".git/hooks/commit-msg", "755");
 	});
-
+	
 	grunt.registerTask("docBuild", ["copy:doc", "clean:doc", "jsdoc"]);
 	grunt.registerTask("build", ["validate-commit", "concat", "uglify", "clean:pkgd", "docBuild"]);
 	grunt.registerTask("default", ["jshint", "jscs", "build", "test"]);
